add content helpers; cleanup
